chore: change name exception for logger
