--- src/common/data/prompt-exception-props.ts
@@ -1,10 +1,0 @@
-export interface PromptExeptionProps {
-  message: string
-  interruptProcess?: boolean
-  exceptionType: exceptionType
-}
-
-export enum exceptionType {
-  WARN = 'WARN',
-  ERROR = 'ERROR',
-}
